Add in-jvm dtest resource leak reproducer. Improve cleanup.
Add a test for repeatedly creating/tearing down in-JVM dtest
clusters to help find resource leaks.

Change IsolatedExecutor to clean up on an executor with zero
core threads so that it exits sooner enabling GC on the
InstanceClassLoader.

Added a check for shutdown in the OutboundTcpConnection connect()
loop, otherwise the thread has to time out before it will exit.

Update the logback config for dtests to make it obvious
which log threads are for instances vs the main logger.