Correctly re-serialize pre-4.0 messages with per-callback serializers
When receiving a message, 4.0 dtest instance has to re-serialize
already decoded messages to pass them through the dtest message filtering.

The 4.0 internode messaging refactor does not expect to have to re-serialize response
messages, and until this patch does not look for message callback serializers for pre-4.0
messages and subsqeuently NPEs.

The reverted previous patch from CASSANDRA-15946 worked because the only messages
that have their payload examined are ReadResponses.

This patch generally fixes the issue by looking up the message callback serializers
for pre-4.0 messages when needed.  The performance impact is expected to be negligible
and only present for connections to pre-4.0 nodes.

Patch by Jon Meredith for CASSANDRA-?; Reviewed by ?.